requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
pandas>=2.0.0
